+++ app/types/ForecastOneDay.ts
