+++ app/utils/weatherDescriptions.ts
