+++ app/utils/weatherGradients.ts
